Adding Subtask service with its impl's

+++ src/main/java/com/SpringProj/todo/Services/Subtask/SubtaskServiceImpl.java
@@ -9,8 +9,10 @@ import com.SpringProj.todo.Model.User;
 import com.SpringProj.todo.Repository.SubTaskRepository;
 import com.SpringProj.todo.Repository.TaskRepository;
 import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,16 +20,19 @@ import java.util.List;
 @Service
 @Data
 @RequiredArgsConstructor
-
-public class SubtaskServiceImpl {
+@Transactional
+public class SubtaskServiceImpl implements SubtaskService {
 
     private final SubTaskRepository subTaskRepository;
     private final TaskRepository taskRepository;
 
-    public SubTask getById(Long id)
+    public SubTask getById(Long id, User user)
     {
+        if(checkAuthorized(id, user))
         return subTaskRepository.findById(id).orElseThrow(() ->
                 new EntityNotFoundException("Subtask not found"));
+
+        else throw new AccessDeniedException("Access denied");
     }
 
 
@@ -54,8 +59,12 @@ public class SubtaskServiceImpl {
         return subTaskRepository.save(subTask);
     }
 
-    public SubTask updateSubTask(Long id, SubTaskUpdateDto subTaskUpdateDto)
+    public SubTask updateSubTask(Long id, SubTaskUpdateDto subTaskUpdateDto, User user)
     {
+        if(!checkAuthorized(id, user))
+            throw new AccessDeniedException("Access denied");
+
+
         Task task = taskRepository.findById(subTaskUpdateDto.getTaskId()).orElseThrow(() ->
                 new EntityNotFoundException("TaskId is not valid"));
 
@@ -70,11 +79,45 @@ public class SubtaskServiceImpl {
         return subTaskRepository.save(subTask);
     }
 
-    public void deleteSubTask(Long id)
+    public void deleteSubTask(Long id, User user)
     {
+        if(!checkAuthorized(id, user))
+            throw new AccessDeniedException("Access denied");
+
+
         SubTask subTask = subTaskRepository.findById(id).orElseThrow(() ->
                 new EntityNotFoundException("Subtask is not found"));
 
         subTaskRepository.delete(subTask);
     }
+
+    public SubTask ChangeStatus(Long id, String status, User user)
+    {
+        SubTask subTask = subTaskRepository.findById(id).orElseThrow(() ->
+                new EntityNotFoundException("Subtask is not found"));
+
+        Task task = taskRepository.findById(subTask.getTask().getId()).orElseThrow(() ->
+                new EntityNotFoundException("Task is not found"));
+
+        if(task.getUser() != user)
+            throw new AccessDeniedException("Access denied");
+
+
+        subTask.setStatus(TaskStatus.valueOf(status));
+
+        return subTaskRepository.save(subTask);
+    }
+
+    private boolean checkAuthorized(Long id, User user)
+    {
+        SubTask subTask = subTaskRepository.findById(id).orElseThrow(() ->
+                new EntityNotFoundException("Subtask is not found"));
+
+        if(subTask.getTask().getUser() == user)
+            return true;
+
+        return false;
+    }
+
+
 }
